requirements.txt
blinker==1.9.0
click==8.3.0
coloredlogs==15.0.1
decorator==5.2.1
Flask==3.1.2
flatbuffers==25.9.23
humanfriendly==10.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
moviepy==2.2.1
mpmath==1.3.0
numpy==2.3.3
onnxruntime==1.23.0
packaging==25.0
pillow==11.3.0
piper-tts==1.3.0
proglog==0.1.12
protobuf==6.32.1
python-dotenv==1.1.1
sympy==1.14.0
tqdm==4.67.1
Werkzeug==3.1.3
